requirements.txt
argon2-cffi==19.2.0
astroid==2.3.3
cffi==1.13.2
Click==7.0
Flask==1.1.1
Flask-JWT-Extended==3.24.1
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.4
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.11
Werkzeug==0.16.0
wrapt==1.11.2
